Remove/simplify Spreadsheet class
